requirements.txt
aiohttp==3.8.5
aiosignal==1.3.1
altair==5.0.1
async-timeout==4.0.2
attrs==23.1.0
azure-core==1.31.0
azure-identity==1.19.0
azure-keyvault-secrets==4.8.0
blinker==1.6.2
build==1.2.2.post1
CacheControl==0.14.0
cachetools==5.3.1
certifi==2023.5.7
cffi==1.17.1
charset-normalizer==3.2.0
cleo==2.1.0
click==8.1.5
crashtest==0.4.1
cryptography==43.0.1
dataclasses-json==0.5.9
decorator==5.1.1
distlib==0.3.9
dulwich==0.21.7
fastjsonschema==2.20.0
filelock==3.16.1
frozenlist==1.4.0
gitdb==4.0.10
GitPython==3.1.32
greenlet==2.0.2
html5lib==1.1
idna==3.4
importlib-metadata==6.8.0
iniconfig==2.0.0
installer==0.7.0
isodate==0.7.2
jaraco.classes==3.4.0
jeepney==0.8.0
Jinja2==3.1.2
jsonschema==4.18.3
jsonschema-specifications==2023.6.1
keyring==24.3.1
langchain==0.0.237
langsmith==0.0.10
lockfile==0.12.2
markdown-it-py==3.0.0
MarkupSafe==2.1.3
marshmallow==3.19.0
marshmallow-enum==1.5.1
mdurl==0.1.2
more-itertools==10.5.0
msal==1.31.0
msal-extensions==1.2.0
msgpack==1.1.0
multidict==6.0.4
mypy-extensions==1.0.0
numexpr==2.8.4
numpy==1.25.1
openai==0.27.8
openapi-schema-pydantic==1.2.4
packaging==23.1
pandas==2.0.3
pexpect==4.9.0
Pillow==9.5.0
pkginfo==1.11.2
platformdirs==3.11.0
pluggy==1.2.0
poetry==1.8.3
poetry-core==1.9.0
poetry-plugin-export==1.8.0
portalocker==2.10.1
protobuf==4.23.4
ptyprocess==0.7.0
pyarrow==12.0.1
pycparser==2.22
pydantic==1.10.11
pydeck==0.8.0
Pygments==2.15.1
PyJWT==2.9.0
Pympler==1.0.1
pyproject_hooks==1.2.0
pytest==7.4.0
pytest-mock==3.11.1
python-dateutil==2.8.2
pytz==2023.3
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0.1
RapidFuzz==3.10.0
referencing==0.29.1
regex==2023.6.3
requests==2.31.0
requests-toolbelt==1.0.0
rich==13.4.2
rpds-py==0.8.11
SecretStorage==3.3.3
shellingham==1.5.4
six==1.16.0
smmap==5.0.0
SQLAlchemy==2.0.19
streamlit==1.24.1
tenacity==8.2.2
tiktoken==0.4.0
toml==0.10.2
tomlkit==0.13.2
toolz==0.12.0
tornado==6.3.2
tqdm==4.65.0
trove-classifiers==2024.10.13
typing-inspect==0.9.0
typing_extensions==4.7.1
tzdata==2023.3
tzlocal==4.3.1
urllib3==2.0.3
validators==0.20.0
virtualenv==20.26.6
watchdog==3.0.0
webencodings==0.5.1
yarl==1.9.2
zipp==3.16.2
